+++ src/Client/Grpc/GrpcDurableTaskClient.cs
@@ -109,7 +109,7 @@ public sealed class GrpcDurableTaskClient : DurableTaskClient
     }
 
     /// <inheritdoc/>
-    public override async Task TerminateAsync(
+    public override async Task TerminateInstanceAsync(
         string instanceId, object? output = null, CancellationToken cancellation = default)
     {
         Check.NotNullOrEmpty(instanceId);
@@ -137,8 +137,7 @@ public sealed class GrpcDurableTaskClient : DurableTaskClient
 
         try
         {
-            await this.sidecarClient.SuspendInstanceAsync(
-                request, cancellationToken: cancellation);
+            await this.sidecarClient.SuspendInstanceAsync(request, cancellationToken: cancellation);
         }
         catch (RpcException e) when (e.StatusCode == StatusCode.Cancelled)
         {
@@ -159,8 +158,7 @@ public sealed class GrpcDurableTaskClient : DurableTaskClient
 
         try
         {
-            await this.sidecarClient.ResumeInstanceAsync(
-                request, cancellationToken: cancellation);
+            await this.sidecarClient.ResumeInstanceAsync(request, cancellationToken: cancellation);
         }
         catch (RpcException e) when (e.StatusCode == StatusCode.Cancelled)
         {
@@ -170,7 +168,7 @@ public sealed class GrpcDurableTaskClient : DurableTaskClient
     }
 
     /// <inheritdoc/>
-    public override async Task<OrchestrationMetadata?> GetInstanceMetadataAsync(
+    public override async Task<OrchestrationMetadata?> GetInstancesAsync(
         string instanceId, bool getInputsAndOutputs = false, CancellationToken cancellation = default)
     {
         if (string.IsNullOrEmpty(instanceId))
@@ -196,7 +194,7 @@ public sealed class GrpcDurableTaskClient : DurableTaskClient
     }
 
     /// <inheritdoc/>
-    public override AsyncPageable<OrchestrationMetadata> GetInstances(OrchestrationQuery? query = null)
+    public override AsyncPageable<OrchestrationMetadata> GetAllInstancesAsync(OrchestrationQuery? filter = null)
     {
         return Pageable.Create(async (continuation, pageSize, cancellation) =>
         {
@@ -204,23 +202,23 @@ public sealed class GrpcDurableTaskClient : DurableTaskClient
             {
                 Query = new P.InstanceQuery
                 {
-                    CreatedTimeFrom = query?.CreatedFrom?.ToTimestamp(),
-                    CreatedTimeTo = query?.CreatedTo?.ToTimestamp(),
-                    FetchInputsAndOutputs = query?.FetchInputsAndOutputs ?? false,
-                    InstanceIdPrefix = query?.InstanceIdPrefix,
-                    MaxInstanceCount = pageSize ?? query?.PageSize ?? OrchestrationQuery.DefaultPageSize,
-                    ContinuationToken = continuation ?? query?.ContinuationToken,
+                    CreatedTimeFrom = filter?.CreatedFrom?.ToTimestamp(),
+                    CreatedTimeTo = filter?.CreatedTo?.ToTimestamp(),
+                    FetchInputsAndOutputs = filter?.FetchInputsAndOutputs ?? false,
+                    InstanceIdPrefix = filter?.InstanceIdPrefix,
+                    MaxInstanceCount = pageSize ?? filter?.PageSize ?? OrchestrationQuery.DefaultPageSize,
+                    ContinuationToken = continuation ?? filter?.ContinuationToken,
                 },
             };
 
-            if (query?.Statuses is not null)
+            if (filter?.Statuses is not null)
             {
-                request.Query.RuntimeStatus.AddRange(query.Statuses.Select(x => x.ToGrpcStatus()));
+                request.Query.RuntimeStatus.AddRange(filter.Statuses.Select(x => x.ToGrpcStatus()));
             }
 
-            if (query?.TaskHubNames is not null)
+            if (filter?.TaskHubNames is not null)
             {
-                request.Query.TaskHubNames.AddRange(query.TaskHubNames);
+                request.Query.TaskHubNames.AddRange(filter.TaskHubNames);
             }
 
             try
@@ -228,7 +226,7 @@ public sealed class GrpcDurableTaskClient : DurableTaskClient
                 P.QueryInstancesResponse response = await this.sidecarClient.QueryInstancesAsync(
                     request, cancellationToken: cancellation);
 
-                bool getInputsAndOutputs = query?.FetchInputsAndOutputs ?? false;
+                bool getInputsAndOutputs = filter?.FetchInputsAndOutputs ?? false;
                 IReadOnlyList<OrchestrationMetadata> values = response.OrchestrationState
                     .Select(x => this.CreateMetadata(x, getInputsAndOutputs))
                     .ToList();
@@ -238,7 +236,7 @@ public sealed class GrpcDurableTaskClient : DurableTaskClient
             catch (RpcException e) when (e.StatusCode == StatusCode.Cancelled)
             {
                 throw new OperationCanceledException(
-                    $"The {nameof(this.GetInstances)} operation was canceled.", e, cancellation);
+                    $"The {nameof(this.GetInstancesAsync)} operation was canceled.", e, cancellation);
             }
         });
     }
@@ -294,7 +292,7 @@ public sealed class GrpcDurableTaskClient : DurableTaskClient
     }
 
     /// <inheritdoc/>
-    public override Task<PurgeResult> PurgeInstanceMetadataAsync(
+    public override Task<PurgeResult> PurgeInstanceAsync(
         string instanceId, CancellationToken cancellation = default)
     {
         this.logger.PurgingInstanceMetadata(instanceId);
@@ -304,7 +302,7 @@ public sealed class GrpcDurableTaskClient : DurableTaskClient
     }
 
     /// <inheritdoc/>
-    public override Task<PurgeResult> PurgeInstancesAsync(
+    public override Task<PurgeResult> PurgeAllInstancesAsync(
         PurgeInstancesFilter filter, CancellationToken cancellation = default)
     {
         this.logger.PurgingInstances(filter);
@@ -374,7 +372,7 @@ public sealed class GrpcDurableTaskClient : DurableTaskClient
         catch (RpcException e) when (e.StatusCode == StatusCode.Cancelled)
         {
             throw new OperationCanceledException(
-                $"The {nameof(this.PurgeInstancesAsync)} operation was canceled.", e, cancellation);
+                $"The {nameof(this.PurgeAllInstancesAsync)} operation was canceled.", e, cancellation);
         }
     }